pyproject.toml
[tool.poetry]
authors = [
    "Rhys Harden <hardenr02@student.psdr3.org>",
    "Victor LoParco <117766368+LoParco1@users.noreply.github.com>",
]
description = "Street Fighter + Chess + (Among) Us. Just the best stuff."
name = "CSP Fighter"
packages = [
    { include = "src/csp-fighter" },
]
version = "0.1.0"

[tool.poetry.dependencies]
pygame = "^2.1.2"
python = ">=3.8.0,<3.11"
pillow = "^9.4.0"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
python-lsp-server = { extras = [
    "pyflakes",
    "rope",
    "yapf",
], version = "^1.5.0" }
toml = "^0.10.2"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.4"
pydocstyle = "^6.3.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.black]
line-length = 88
preview = false


[tool.isort]
atomic = true
line_length = 88
profile = "black"
skip_gitignore = true
